feat: implement support for frigidaire gallery inverter window room air conditioner
