+++ package.json
@@ -1,7 +1,7 @@
 {
     "displayName": "Homebridge Electrolux Devices",
     "name": "homebridge-electrolux-devices",
-    "version": "1.1.0",
+    "version": "1.2.0-beta.1",
     "description": "Homebridge plugin for Electrolux devices",
     "license": "Apache-2.0",
     "repository": {
@@ -23,7 +23,9 @@
         "watch": "npm run build && npm link && nodemon"
     },
     "keywords": [
-        "homebridge-plugin"
+        "homebridge-plugin",
+        "electrolux",
+        "frigidaire"
     ],
     "dependencies": {
         "axios": "^1.4.0",
